blogpost_backend/pyproject.toml
[project]
name = "blogpost-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.119.0",
    "google-generativeai>=0.8.5",
    "pymongo>=4.15.3",
    "uvicorn>=0.37.0",
]
